updated to v1.10.2

+++ Tool/EraserTool.cs
@@ -15,12 +15,12 @@ namespace Vintagestory.ServerMods.WorldEdit
             
         }
 
-        public override bool ApplyToolBuild(WorldEdit worldEdit, Block placedBlock, ushort oldBlockId, BlockSelection blockSel, BlockPos targetPos, ItemStack withItemStack)
+        public override bool ApplyToolBuild(WorldEdit worldEdit, Block placedBlock, int oldBlockId, BlockSelection blockSel, BlockPos targetPos, ItemStack withItemStack)
         {
             return false;
         }
 
-        public override bool ApplyToolBreak(WorldEdit worldEdit, Block block, ushort oldBlockId, BlockSelection blockSel, BlockPos targetPos)
+        public override bool ApplyToolBreak(WorldEdit worldEdit, Block block, int oldBlockId, BlockSelection blockSel, BlockPos targetPos)
         {
             block = blockAccessRev.GetBlock(0);
             return PerformBrushAction(worldEdit, block, oldBlockId, blockSel, targetPos, null);
