+++ Tool/ImportTool.cs
@@ -302,12 +302,12 @@ namespace Vintagestory.ServerMods.WorldEdit
             return false;
         }
 
-        public override void OnBreak(WorldEdit worldEdit, ushort oldBlockId, BlockSelection blockSel)
+        public override void OnBreak(WorldEdit worldEdit, int oldBlockId, BlockSelection blockSel)
         {
             
         }
 
-        public override void OnBuild(WorldEdit worldEdit, ushort oldBlockId, BlockSelection blockSel, ItemStack withItemStack)
+        public override void OnBuild(WorldEdit worldEdit, int oldBlockId, BlockSelection blockSel, ItemStack withItemStack)
         {
             if (blockDatas == null)
             {
