+++ Tool/RaiseLowerTool.cs
@@ -118,18 +118,18 @@ namespace Vintagestory.ServerMods.WorldEdit
         }
 
        
-        public override void OnBreak(WorldEdit worldEdit, ushort oldBlockId, BlockSelection blockSel)
+        public override void OnBreak(WorldEdit worldEdit, int oldBlockId, BlockSelection blockSel)
         {
             OnUse(worldEdit, blockSel.Position, oldBlockId, -1, null);
         }
 
-        public override void OnBuild(WorldEdit worldEdit, ushort oldBlockId, BlockSelection blockSel, ItemStack withItemStack)
+        public override void OnBuild(WorldEdit worldEdit, int oldBlockId, BlockSelection blockSel, ItemStack withItemStack)
         {
             OnUse(worldEdit, blockSel.Position, oldBlockId, 1, withItemStack);
         }
 
 
-        void OnUse(WorldEdit worldEdit, BlockPos pos, ushort oldBlockId, int sign, ItemStack withItemStack)
+        void OnUse(WorldEdit worldEdit, BlockPos pos, int oldBlockId, int sign, ItemStack withItemStack)
         {
             if (Radius <= 0) return;
